added message size check to MAC block
the MAC block now calculates the maximal number of encoded
Bits that can be sent within one data slot.
if a "packet" can't be sent within one slot, it is dropped.

information about the exact transmission timing and the maximum
number of bits to send (= max bits in data + location slot) is
passed to the lower layer using the metadata dictionary
of the PDU message